balena: Fix dnsmasq dependency
dnsmasq should run before balena daemons to ensure that follow up services have
a chance of pulling images.

This is useful when provisioning a device with supervisor not included in the
OS. Ex.: Running resinOS in a container. In this case, connectivity is assumed
and at boot update-supervisor will try to pull supervisor. If dns is not
available it will fail and will be retriggered when update supervisor timer
unit will get activated which delays the boot.

Signed-off-by: Andrei Gherzan <andrei@resin.io>