update-resin-supervisor: Don't fail if there is no DEVICE ID
The device id is generated when registering the device, which is done by the
supervisor. On OS instances where supervisor is not preloaded,
update-supervisor brings in the supervisor at first boot / network
connectivity but this breaks if this service fails when DEVICE ID is not
available. This means that the above exampled resinOS instance, will never
pull supervisor because it expects something supervisor is generating in the
first place.

This change will not impact "unregistered devices with not connectivity but
supervisor pre loaded" either because the script checks later on if supervisor
is already pulled and exits successfully.

Signed-off-by: Andrei Gherzan <andrei@resin.io>